MODUIMP-50: Add "tags" property to userdataimport.json schema
Copy
https://github.com/folio-org/mod-users/blob/v18.0.0/ramls/userdata.json
to ramls/schemas/userdataimport.json
and add "requestPreference" and "required" properties.

This fixes the missing "tags" property that was added to userdata.json
but is missing from userdataimport.json.

Move submodule from ramls/raml-util to ramls/schemas/raml-util
so that the paths are the same as in userdata.json.